Trademark updates
I'll send you the necessary footer by email. I tried to change to Iso-Seq wherever possible. In the table under Subset Options, in the Example column, there are two names that have "isoseq" in them, but I think those are file names that are generated by the instrument so can't be changed. Also, in the table under Diff SMRT Analysis, the bottom cells have isoseq in them and I don't know if we can have those say Iso-Seq or not. Thanks!